Moved everything to battle_move_effects.h
